+++ tests/execution_engine/split_and_sample/test_pandas_execution_engine_splitting.py
@@ -1,3 +1,5 @@
+import datetime
+import os
 from typing import List
 from unittest import mock
 
@@ -5,8 +7,16 @@ import pandas as pd
 import pandas.api.types as ptypes
 import pytest
 
-from great_expectations.core.batch_spec import RuntimeDataBatchSpec
-from great_expectations.execution_engine import PandasExecutionEngine
+import great_expectations.exceptions as ge_exceptions
+from great_expectations.core import IDDict
+from great_expectations.core.batch import BatchDefinition
+from great_expectations.core.batch_spec import (
+    PathBatchSpec,
+    RuntimeDataBatchSpec,
+    S3BatchSpec,
+)
+from great_expectations.datasource.data_connector import ConfiguredAssetS3DataConnector
+from great_expectations.execution_engine import ExecutionEngine, PandasExecutionEngine
 from great_expectations.execution_engine.split_and_sample.data_splitter import DatePart
 from great_expectations.execution_engine.split_and_sample.pandas_data_splitter import (
     PandasDataSplitter,
@@ -35,7 +45,6 @@ MULTIPLE_DATE_PART_DATE_PARTS += [
 
 @pytest.fixture
 def simple_multi_year_pandas_df():
-
     df: pd.DataFrame = pd.DataFrame(
         data={
             "input_timestamp": [
@@ -57,6 +66,20 @@ def simple_multi_year_pandas_df():
 
     assert len(df.index) == 9
     return df
+
+
+@pytest.fixture
+def batch_with_split_on_whole_table_s3(test_s3_files) -> S3BatchSpec:
+    bucket, keys = test_s3_files
+    path = keys[0]
+    full_path = f"s3a://{os.path.join(bucket, path)}"
+
+    batch_spec = S3BatchSpec(
+        path=full_path,
+        reader_method="read_csv",
+        splitter_method="_split_on_whole_table",
+    )
+    return batch_spec
 
 
 @pytest.mark.integration
@@ -257,3 +280,234 @@ def test_get_splitter_method(underscore_prefix: str, splitter_method_name: str):
     assert data_splitter.get_splitter_method(
         splitter_method_name_with_prefix
     ) == getattr(data_splitter, splitter_method_name)
+
+
+def test_get_batch_with_split_on_whole_table_runtime(test_df):
+    split_df = PandasExecutionEngine().get_batch_data(
+        RuntimeDataBatchSpec(
+            batch_data=test_df, splitter_method="_split_on_whole_table"
+        )
+    )
+    assert split_df.dataframe.shape == (120, 10)
+
+
+def test_get_batch_with_split_on_whole_table_filesystem(
+    test_folder_connection_path_csv,
+):
+    test_df = PandasExecutionEngine().get_batch_data(
+        PathBatchSpec(
+            path=os.path.join(test_folder_connection_path_csv, "test.csv"),
+            reader_method="read_csv",
+            splitter_method="_split_on_whole_table",
+        )
+    )
+    assert test_df.dataframe.shape == (5, 2)
+
+
+def test_get_batch_with_split_on_whole_table_s3(
+    batch_with_split_on_whole_table_s3, test_df_small
+):
+    df = PandasExecutionEngine().get_batch_data(
+        batch_spec=batch_with_split_on_whole_table_s3
+    )
+    assert df.dataframe.shape == test_df_small.shape
+
+
+def test_get_batch_with_split_on_whole_table_s3_with_configured_asset_s3_data_connector(
+    test_s3_files, test_df_small
+):
+    bucket, _keys = test_s3_files
+    expected_df = test_df_small
+
+    execution_engine: ExecutionEngine = PandasExecutionEngine()
+
+    my_data_connector = ConfiguredAssetS3DataConnector(
+        name="my_data_connector",
+        datasource_name="FAKE_DATASOURCE_NAME",
+        bucket=bucket,
+        execution_engine=execution_engine,
+        prefix="",
+        assets={"alpha": {}},
+        default_regex={
+            "pattern": "alpha-(.*)\\.csv",
+            "group_names": ["index"],
+        },
+    )
+    batch_def: BatchDefinition = BatchDefinition(
+        datasource_name="FAKE_DATASOURCE_NAME",
+        data_connector_name="my_data_connector",
+        data_asset_name="alpha",
+        batch_identifiers=IDDict(index=1),
+        batch_spec_passthrough={
+            "reader_method": "read_csv",
+            "splitter_method": "_split_on_whole_table",
+        },
+    )
+    test_df = execution_engine.get_batch_data(
+        batch_spec=my_data_connector.build_batch_spec(batch_definition=batch_def)
+    )
+    assert test_df.dataframe.shape == expected_df.shape
+
+    # if key does not exist
+    batch_def_no_key = BatchDefinition(
+        datasource_name="FAKE_DATASOURCE_NAME",
+        data_connector_name="my_data_connector",
+        data_asset_name="alpha",
+        batch_identifiers=IDDict(index=9),
+        batch_spec_passthrough={
+            "reader_method": "read_csv",
+            "splitter_method": "_split_on_whole_table",
+        },
+    )
+    with pytest.raises(ge_exceptions.ExecutionEngineError):
+        execution_engine.get_batch_data(
+            batch_spec=my_data_connector.build_batch_spec(
+                batch_definition=batch_def_no_key
+            )
+        )
+
+
+def test_get_batch_with_split_on_column_value(test_df):
+    split_df = PandasExecutionEngine().get_batch_data(
+        RuntimeDataBatchSpec(
+            batch_data=test_df,
+            splitter_method="_split_on_column_value",
+            splitter_kwargs={
+                "column_name": "batch_id",
+                "batch_identifiers": {"batch_id": 2},
+            },
+        )
+    )
+    assert split_df.dataframe.shape == (12, 10)
+    assert (split_df.dataframe.batch_id == 2).all()
+
+    split_df = PandasExecutionEngine().get_batch_data(
+        RuntimeDataBatchSpec(
+            batch_data=test_df,
+            splitter_method="_split_on_column_value",
+            splitter_kwargs={
+                "column_name": "date",
+                "batch_identifiers": {"date": datetime.date(2020, 1, 30)},
+            },
+        )
+    )
+    assert split_df.dataframe.shape == (3, 10)
+
+
+def test_get_batch_with_split_on_converted_datetime(test_df):
+    split_df = PandasExecutionEngine().get_batch_data(
+        RuntimeDataBatchSpec(
+            batch_data=test_df,
+            splitter_method="_split_on_converted_datetime",
+            splitter_kwargs={
+                "column_name": "timestamp",
+                "batch_identifiers": {"timestamp": "2020-01-30"},
+            },
+        )
+    )
+    assert split_df.dataframe.shape == (3, 10)
+
+
+def test_get_batch_with_split_on_divided_integer(test_df):
+    split_df = PandasExecutionEngine().get_batch_data(
+        RuntimeDataBatchSpec(
+            batch_data=test_df,
+            splitter_method="_split_on_divided_integer",
+            splitter_kwargs={
+                "column_name": "id",
+                "divisor": 10,
+                "batch_identifiers": {"id": 5},
+            },
+        )
+    )
+    assert split_df.dataframe.shape == (10, 10)
+    assert split_df.dataframe.id.min() == 50
+    assert split_df.dataframe.id.max() == 59
+
+
+def test_get_batch_with_split_on_mod_integer(test_df):
+    split_df = PandasExecutionEngine().get_batch_data(
+        RuntimeDataBatchSpec(
+            batch_data=test_df,
+            splitter_method="_split_on_mod_integer",
+            splitter_kwargs={
+                "column_name": "id",
+                "mod": 10,
+                "batch_identifiers": {"id": 5},
+            },
+        )
+    )
+    assert split_df.dataframe.shape == (12, 10)
+    assert split_df.dataframe.id.min() == 5
+    assert split_df.dataframe.id.max() == 115
+
+
+def test_get_batch_with_split_on_multi_column_values(test_df):
+    split_df = PandasExecutionEngine().get_batch_data(
+        RuntimeDataBatchSpec(
+            batch_data=test_df,
+            splitter_method="_split_on_multi_column_values",
+            splitter_kwargs={
+                "column_names": ["y", "m", "d"],
+                "batch_identifiers": {
+                    "y": 2020,
+                    "m": 1,
+                    "d": 5,
+                },
+            },
+        )
+    )
+    assert split_df.dataframe.shape == (4, 10)
+    assert (split_df.dataframe.date == datetime.date(2020, 1, 5)).all()
+
+    with pytest.raises(ValueError):
+        # noinspection PyUnusedLocal
+        split_df = PandasExecutionEngine().get_batch_data(
+            RuntimeDataBatchSpec(
+                batch_data=test_df,
+                splitter_method="_split_on_multi_column_values",
+                splitter_kwargs={
+                    "column_names": ["I", "dont", "exist"],
+                    "batch_identifiers": {
+                        "y": 2020,
+                        "m": 1,
+                        "d": 5,
+                    },
+                },
+            )
+        )
+
+
+def test_get_batch_with_split_on_hashed_column(test_df):
+    with pytest.raises(ge_exceptions.ExecutionEngineError):
+        # noinspection PyUnusedLocal
+        split_df = PandasExecutionEngine().get_batch_data(
+            RuntimeDataBatchSpec(
+                batch_data=test_df,
+                splitter_method="_split_on_hashed_column",
+                splitter_kwargs={
+                    "column_name": "favorite_color",
+                    "hash_digits": 1,
+                    "batch_identifiers": {
+                        "hash_value": "a",
+                    },
+                    "hash_function_name": "I_am_not_valid",
+                },
+            )
+        )
+
+    split_df = PandasExecutionEngine().get_batch_data(
+        RuntimeDataBatchSpec(
+            batch_data=test_df,
+            splitter_method="_split_on_hashed_column",
+            splitter_kwargs={
+                "column_name": "favorite_color",
+                "hash_digits": 1,
+                "batch_identifiers": {
+                    "hash_value": "a",
+                },
+                "hash_function_name": "sha256",
+            },
+        )
+    )
+    assert split_df.dataframe.shape == (8, 10)
